move highlight component logic to config/utils
