+++ src/config/styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native'
+
+export default styleUtils = StyleSheet.create({
+  btnIcon: { 
+    backgroundColor: "#fff", 
+    borderWidth: 0
+  },
+  searchResult: { 
+    color: "#4a148c", 
+    backgroundColor: "yellow",
+    fontWeight: "bold",
+  },
+});
