+++ src/config/utils.js
@@ -1,4 +1,7 @@
+import React from 'react'
 import VersionNumber from 'react-native-version-number'
+import styleUtils from './styles'
+import { Text } from 'native-base'
 
 const typeIcon = (docType) => {
   const type = 'MaterialIcons'
@@ -19,4 +22,24 @@ const realname = (path) => {
 
 const appVersion = VersionNumber.appVersion
 
-export { typeIcon, basename, realname, appVersion }
+const highlight = (text, Tag) => {
+  let data = []
+  words = text.split(/\s/)
+  let regex = /class='highlight'>\w+<\/em>/
+
+  words.forEach( (word, index) => {
+    if( word == '<em' && regex.test(words[index+1]) ) return
+
+    if( regex.test(word) ) {
+      hl = />(\w+)</.exec(word)
+      data.push(<Tag key={index} style={styleUtils.searchResult}>{hl[1]}</Tag>)
+    } else  {
+      data.push(<Tag key={index}>{word}</Tag>)
+    }
+    data.push(<Text key={index + Date.now()}>{' '}</Text>)
+  })
+
+  return data
+}
+
+export { typeIcon, basename, realname, appVersion, highlight }
